✏️ Fix: database schema

+++ src/main/java/com/aloc/aloc/domain/solvedproblem/SolvedProblem.java
@@ -19,7 +19,7 @@ import lombok.Getter;
 public class SolvedProblem {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id")
@@ -30,8 +30,4 @@ public class SolvedProblem {
 	private Problem problem;
 
 	private LocalDateTime solvedAt;
-
-
-
-
 }
